requirements.txt
aiohttp>=3.11.13
pandas>=2.0.3
python-dotenv>=1.0.0
tqdm>=4.67.1
pydantic>=2.10.6
pydantic-ai==0.0.36
openai>=1.65.4
beautifulsoup4>=4.13.3
requests>=2.32.3
tenacity>=8.2.3